refactor(subprocess): split OS runtime modules (#0000)
